311: address PR review — module renames, scope cleanup, IIR fix
- Rename `wies/core/permissions.py` → `permission_engine.py` and
  `wies/core/permission_rules.py` → `permissions.py`. Reviewer expected
  application-specific permissions in `permissions.py` per Django
  convention; the engine moves out of the way under a more accurate name.
- Drop `LIST` and `CREATE` from the `Verb` enum. The engine is row-level
  only; class-level access stays on Django decorators.
- Rename `wies/core/editables/user_profile.py` → `user.py` to match
  the model name.
- Tighten `update_assignment_extra_info` so externally-managed
  assignments (IIR) never show a pencil, even for placed colleagues.
- Fix the dev-user docstring in `setup_initial_user` — group membership,
  not superuser status, is what gives the dev user edit rights.
- Add an audit `create_event` for assignment creation in
  `assignment_create` so it appears on the Updates tab.
- Explain the F405 noqa in `production.py`.
- Compress the 311 changelog entries into a single user-facing line.
