+++ config/settings/production.py
@@ -38,6 +38,9 @@ STORAGES = {
 # WhiteNoise serves static files. Insert it directly after SecurityMiddleware
 # per the WhiteNoise docs — kept out of the base MIDDLEWARE list so
 # tests/local don't need to filter it out.
+# F405 (defined-from-star-import): MIDDLEWARE comes from `from .base import *`
+# above; ruff can't see it. Safe to skip — this file is the only place
+# where the star-imported names are mutated.
 _WHITENOISE = "whitenoise.middleware.WhiteNoiseMiddleware"
 _AFTER_SECURITY = MIDDLEWARE.index("django.middleware.security.SecurityMiddleware") + 1  # noqa: F405
 MIDDLEWARE.insert(_AFTER_SECURITY, _WHITENOISE)  # noqa: F405
